refactor: model.push -> model.attendance
