+++ app/domain/model/attendance.py
